api/python-validate/requirements.txt
guardrails-ai==0.6.6
pydantic>=2.0.0
requests>=2.28.0
nltk>=3.8.0
transformers>=4.21.0
torch>=1.12.0
numpy>=1.21.0 